requirements.txt
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
dnspython==2.3.0
exceptiongroup==1.1.1
Flask==2.3.2
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
load-dotenv==0.1.0
MarkupSafe==2.1.2
packaging==23.1
pluggy==1.0.0
pymongo==4.3.3
pytest==7.3.1
pytest-mock==3.10.0
python-dotenv==1.0.0
PyYAML==6.0
requests==2.30.0
six==1.16.0
soupsieve==2.4.1
tomli==2.0.1
types-PyYAML==6.0.12.10
urllib3==2.0.2
Werkzeug==2.3.3
gunicorn==20.1.0